pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zimage-trainer"
version = "0.1.0"
description = "A modular training toolkit for Z-Image (Lumina-Image) LoRA fine-tuning"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "torchaudio>=2.8.0",
    "diffusers>=0.27.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "safetensors>=0.7.0",
    "bitsandbytes>=0.42.0",
    "lpips>=0.1.4",
    "toml>=0.10.0",
    "tqdm>=4.67.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.38.0",
    "pydantic>=2.5.0",
    "websockets>=15.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
zimage-train = "zimage_trainer.train:main"
zimage-cache-latents = "zimage_trainer.cache_latents:main"
zimage-cache-text = "zimage_trainer.cache_text_encoder:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

